Moved components to src directory
