+++ src/components/StoreInfoFetcher/index.ts
